+++ firmware/esp32/main/inter_mcu.h
@@ -34,6 +34,7 @@ typedef void (*wavex_cv_cal_cb_t)(const WaveX::Protocol::CvCalMessage& cal, void
 void inter_mcu_set_cv_cal_listener(wavex_cv_cal_cb_t cb, void* user_data);
 void inter_mcu_invoke_cv_cal_callback(const WaveX::Protocol::CvCalMessage& cal);
 
+esp_err_t inter_mcu_send_midi_program(const WaveX::Protocol::MidiProgramMessage& message);
 esp_err_t inter_mcu_send_midi_clock(const WaveX::Protocol::MidiClockEventMessage& message);
 
 // Basic MIDI message sending